Emit per-evaluation decision snapshot; expose MACD histogram & ADX exit pressure; prioritize emergency exits

+++ src/main/java/com/binance/strategy/StrategyLogV1.java
@@ -29,7 +29,7 @@ public final class StrategyLogV1 {
 			Integer score5m,
 			Integer hamScore,
 			Integer ctiDirScore,
-			Integer macdScore,
+			Double macdScore,
 			Integer scoreLong,
 			Integer scoreShort,
 			Double adjScore,
